+++ test/test_io.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Tests for the iotext module.
+"""Tests for the io module.
 
 BSD 3-Clause License
 Copyright (c) 2019-2020, Daniel Nagel
@@ -11,7 +11,7 @@ import os.path
 import numpy as np
 import pytest
 
-from msmhelper import iotext
+from msmhelper import io
 
 # Current directory
 HERE = os.path.dirname(__file__)
@@ -35,7 +35,7 @@ def test_opentxt(traj_file, kwargs):
     expected = np.array([[1, 2], [1, 3], [1, 2], [1, 1], [1, 2], [2, 1],
                          [2, 5], [1, 4], [2, 3], [3, 2], [2, 1], [2, 2],
                          [3, 3]])
-    data = iotext.opentxt(traj_file, **kwargs)
+    data = io.opentxt(traj_file, **kwargs)
     if len(data.shape) == 1:
         assert (data == expected[:, 0]).all()
     else:
@@ -53,19 +53,19 @@ def test_savetxt(traj_file, header, tmpdir):
     """Test that a file is opened correctly."""
     expected = [1, 1, 1, 1, 1, 2, 2, 1, 2, 3, 2, 2, 3]
     output = tmpdir.join('test_traj.txt')
-    iotext.savetxt(output, expected, header=header, fmt='%.0f')
-    assert (iotext.opentxt(output) == expected).all()
+    io.savetxt(output, expected, header=header, fmt='%.0f')
+    assert (io.opentxt(output) == expected).all()
 
 
 @pytest.mark.parametrize('wrong_limit', [os.path.join(HERE, 'data.dat')])
 def test_open_limits(limits_file, wrong_limit):
     """Test that the limits are loaded correctly."""
-    assert (iotext.open_limits(13, limits_file) == [5, 10, 13]).all()
-    assert (iotext.open_limits(13) == [13]).all()
-    with pytest.raises(iotext.FileError):
-        iotext.open_limits(13, wrong_limit)
+    assert (io.open_limits(13, limits_file) == [5, 10, 13]).all()
+    assert (io.open_limits(13) == [13]).all()
+    with pytest.raises(io.FileError):
+        io.open_limits(13, wrong_limit)
     with pytest.raises(ValueError):
-        iotext.open_limits(14, limits_file)
+        io.open_limits(14, limits_file)
 
 
 @pytest.mark.parametrize('traj_file, kwargs', [
@@ -78,21 +78,21 @@ def test_openmicrostates(limits_file, traj_file, kwargs):
     """Test that the trajectory is split correctly."""
     if 'dtype' in kwargs and kwargs['dtype'] == np.float64:
         with pytest.raises(TypeError):
-            iotext.openmicrostates(
+            io.openmicrostates(
                 limits_file=limits_file,
                 file_name=traj_file,
                 **kwargs,
             )
     elif 'no_traj' in kwargs:
-        with pytest.raises(iotext.FileError):
-            iotext.openmicrostates(
+        with pytest.raises(io.FileError):
+            io.openmicrostates(
                 limits_file=limits_file,
                 file_name=traj_file,
             )
     else:
         # check with limits file
         expected = [[1, 1, 1, 1, 1], [2, 2, 1, 2, 3], [2, 2, 3]]
-        trajs = iotext.openmicrostates(
+        trajs = io.openmicrostates(
             limits_file=limits_file,
             file_name=traj_file,
             **kwargs,
@@ -118,7 +118,7 @@ def test_openmicrostates(limits_file, traj_file, kwargs):
 def test_opentxt_limits(limits_file, traj_file, expected):
     """Test that the trajectory is split correctly."""
     # check with limits file
-    trajs = iotext.opentxt_limits(
+    trajs = io.opentxt_limits(
         limits_file=limits_file,
         file_name=traj_file,
     )
@@ -128,6 +128,6 @@ def test_opentxt_limits(limits_file, traj_file, expected):
     # check without limits file
     if len(trajs[0].shape) == 1:
         expected = [[1, 1, 1, 1, 1, 2, 2, 1, 2, 3, 2, 2, 3]]
-        trajs = iotext.opentxt_limits(file_name=traj_file)
+        trajs = io.opentxt_limits(file_name=traj_file)
         for i, traj in enumerate(trajs):
             assert (traj == expected[i]).all()